feat: renamed web functions and tests
